Created a prototype input scheme which can be used for my game prototype and expanded upon later once I've built stuff up more
scInputDevice: Acts as a forwarding mechanism for keyboard events so
that other things can forward their input into the simulation core
without it having to speak their language (from an object perspective)

KeyboardState: Tracks the state of the keyboard given the input from
scInputDevice

KeyMapping: Tracks which keys are mapped to common actions (up, down,
left, right)

scSimulationObject: an object which needs to be "simulated" throughout
the course of the program. Provides no output by itself.

scControlScheme: An interface which allows an object to be controlled
through some arbitrary means. For instance, through a keyboard or
through AI. This will go hand in hand with scSimulationObject in most
cases, but not necessarily all (i.e. things that talk with CommCore
objects, etc).

scKeyboardControlledObj: Keyboard controlled simulatable object which
moves the position of the thing up/down/left/right based on the
keyboardState and the speed modifier for the class (if, for instance,
one keyboard character needs to go faster than the other). The
positional update mechanics might need to be generified since it's math-

And then probably some other stuff