started commenting using doxygen style, and did basic implementation of a "find adjacent" function.
Didn't do too much on the doxygen commenting front - it's kinda shitty
because now my comments are inconsistant, but it'll take a concerted
effort to get everything, and I need to pick up steam again before I
even think about style.

As for the find adjacent funciton, it's grossely inefficient as it
stands, but it's a start.