Did a bunch of stuff to get cameras "working"
Biggest things: removed coObject, added coWorld, added coCameraManager.
the first was replaced by the second (which manages a bunch of objects
using a relatively clean interface) and the third is what handles
cameraing, using an also relatively clean interface, which is exposed
through coWorld, to a point.

Unfortunately, it doesn't actually work yet. I'm going to have to put in
some debug information to get things working well - a good excuse to get
a debug implementation going. It looks like the math checks out for now,
at a minimum.