Got mothafukin the first resource block test working.
This should include a basic interface for generating chunks of blocks.

You request for it to generate chunks within a certain radius of a
block, and then you can query it for resources using a call back (which
is documented, though not terribly well).

Next is to work on some minor test cases to round out coverage, then
work on filling in the next major feature test.